requirements.txt
click==8.1.3
Flask==2.2.3
graphql-core==3.2.3
graphql-server-core==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
promise==2.3
Rx==1.6.3
six==1.16.0
Werkzeug==2.2.3
